fix tag bug

+++ iOverlay/src/core/link.rs
@@ -10,17 +10,18 @@ pub struct OverlayLink {
     pub b: IdPoint,
     pub fill: SegmentFill,
     /// User-supplied edge tag, propagated from the input `Segment`.
-    pub tag: u32,
+    pub tag: u16,
 }
 
 impl OverlayLink {
     #[inline(always)]
+    #[allow(dead_code)]
     pub(crate) fn new(a: IdPoint, b: IdPoint, fill: SegmentFill) -> OverlayLink {
         OverlayLink { a, b, fill, tag: 0 }
     }
 
     #[inline(always)]
-    pub(crate) fn new_tagged(a: IdPoint, b: IdPoint, fill: SegmentFill, tag: u32) -> OverlayLink {
+    pub(crate) fn new_tagged(a: IdPoint, b: IdPoint, fill: SegmentFill, tag: u16) -> OverlayLink {
         OverlayLink { a, b, fill, tag }
     }
